Send personalised welcome mail to new signups
The welcome mail is send 32 minutes after signup. It also seems to be sent by
the CEO and has no layout. This is supposed to drive engagement nuts with
probably around 17% of the people responding to it.

The mail is send inside an ActiveSupport::Concern. More on that can be found in
app/concerns/welcomed.rb

An automated nicely styled mail should probably be send directly after a user
signes up. This has been stubbed, but since no text/picture material is readily
available, the mail does not get sent.