Add fb live url to venue, adjust userdata creation
I added the db migration and the schema update. Due to a bug/undocumented behavior of ffmpeg with external images as source, we are saving an image during the build of the image.

I adjusted userdata.sh.erb to render the corrent userdata.sh that spawns a ffmpeg container with the fb live url from of the venue.